chkdep.py: added command line options TODO: improved error handling, support for methods other than elf

+++ chkdep.py
@@ -1,6 +1,24 @@
 #!/usr/bin/env python
 
-import tarfile, tempfile, shutil, os, stat, re, pacman
+"""Checks a package or directory for possible depends
+
+Usage: chkdep [options]
+
+Options:
+  -d ..., --dir=...       name od the dir to check
+  -i ..., --ignore=...    ignore the given package (optional, useful when a
+                          package links to itself)
+  -m ..., --method=...    method to use to detect the deps (default: elf,
+                          possible values: elf)
+  -p ..., --package=...   name of the fpm to check
+
+You should at least use the -d or the -p option.
+"""
+
+import tarfile, tempfile, shutil, os, stat, re, pacman, getopt, sys
+
+def usage():
+	print __doc__
 
 def rmdupdeps(deps):
 	depdeps = []
@@ -18,7 +36,7 @@ def rmdupdeps(deps):
 				j = pacman.list_next(j)
 		i = pacman.list_next(i)
 	for i in deps:
-		if i not in depdeps:
+		if i not in depdeps and i not in ignorepkgs:
 			newdeps.append(i)
 	return newdeps
 
@@ -43,23 +61,52 @@ pacman.initialize("/")
 db = pacman.db_register("local")
 ignorepkgs = []
 deps = []
+
+dir = None
+allowedmethods = ("elf")
 method="elf"
+pkg = None
+
+try:
+	opts, args = getopt.getopt(sys.argv[1:], "d:p:i:m:", ["dir=", "package=", "ignore=", "method="])
+except getopt.GetoptError:
+	usage()
+	sys.exit(1)
+
+for opt, arg in opts:
+	if opt in ("-d", "--dir"):
+		dir = arg
+	if opt in ("-p", "--package"):
+		pkg = arg
+	if opt in ("-m", "--method"):
+		method = arg
+	if opt in ("-i", "--ignore"):
+		ignorepkgs.append(arg)
+
+if not dir and not pkg:
+	usage()
+	sys.exit(1)
 
 checker = getattr(checks, method)
-fpmroot = tempfile.mkdtemp()
 
 if "FAKEROOTKEY" in os.environ.keys():
 	ignorepkgs.append("fakeroot")
 
-fpm = tarfile.TarFile.open("pacman-tools-0.8.8-1-i686.fpm", "r:bz2")
-fpm.extractall(fpmroot)
-fpm.close()
+if pkg:
+	fpmroot = tempfile.mkdtemp()
+	fpm = tarfile.TarFile.open(pkg, "r:bz2")
+	fpm.extractall(fpmroot)
+	fpm.close()
+
+if dir:
+	fpmroot = dir
 
 for root, dirs, files in os.walk(fpmroot):
 	for file in files:
 		checker(os.path.join(root, file))
 
-shutil.rmtree(fpmroot)
+if pkg:
+	shutil.rmtree(fpmroot)
 
 deps = rmdupdeps(deps)
 print deps
